refactor(phase-f): split extensions/xp.py into package + features/xp/
Break the 710-line monolithic XP extension into the feature/extension
layering used elsewhere in the codebase.

features/xp/                           — pure logic, no Discord deps
  constants.py   XP/cooldown/leaderboard constants
  leveling.py    calculate_level, get_rank_display, create_progress_bar
  cache.py       generic TTLCache
  repository.py  XpRepository — encapsulates mongo_manager calls for the
                 xp/xp_events collections (ensure_indexes/collection,
                 get_user, insert, update_xp, set_level, log_event,
                 get_user_rank with $setWindowFields + linear fallback,
                 list_all_sorted_by_xp)

extensions/xp/                         — Discord glue, mixin-composed
  _common.py       XpConfig schema + logger + config
  leveling.py      LevelingMixin: on_message listener, cooldown, level-up
  commands.py      CommandsMixin: /rank, /leaderboard
  leaderboard.py   LeaderboardMixin: permanent leaderboard task + embeds
  __init__.py      XpExtension composition + setup()

No behavior change. MongoDB collections ("xp", "xp_events") and config key
(moduleXp) are unchanged, so no migration is needed. The extension still
loads under the existing ``extensions.xp`` entry in main.py's loader —
nothing else imported from the old flat module.
