+++ features/xp/__init__.py
@@ -1,0 +1,32 @@
+"""XP feature — level curve, leaderboard ranking, and MongoDB persistence."""
+
+from features.xp.cache import TTLCache
+from features.xp.constants import (
+    DEFAULT_LEVEL_UP_MESSAGE,
+    LEADERBOARD_PAGE_SIZE,
+    RANK_CACHE_TTL,
+    RANK_MEDALS,
+    USER_CACHE_TTL,
+    XP_COOLDOWN_SECONDS,
+    XP_MAX,
+    XP_MIN,
+)
+from features.xp.leveling import calculate_level, create_progress_bar, get_rank_display
+from features.xp.repository import UserXpStats, XpRepository
+
+__all__ = [
+    "DEFAULT_LEVEL_UP_MESSAGE",
+    "LEADERBOARD_PAGE_SIZE",
+    "RANK_CACHE_TTL",
+    "RANK_MEDALS",
+    "TTLCache",
+    "USER_CACHE_TTL",
+    "UserXpStats",
+    "XP_COOLDOWN_SECONDS",
+    "XP_MAX",
+    "XP_MIN",
+    "XpRepository",
+    "calculate_level",
+    "create_progress_bar",
+    "get_rank_display",
+]
